% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Histogram.R
\name{Histogram}
\alias{Histogram}
\title{Creates a Histogram Plot}
\usage{
Histogram(mod, data, target, var.name)
}
\arguments{
\item{mod}{model object}

\item{data}{\code{data.frame} model object was build from}

\item{target}{name dependent variable}

\item{var.name}{name of independent variable}
}
\value{
Histogram or Bar plot
}
\description{
\code{Historgram} creates a basic histogram plot or bar plot (for categorical varaibles
   and variables with less than 10 values)
}

