% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LorenzCurve.R
\name{Lorenz}
\alias{Lorenz}
\title{Plot Lorenz Curve}
\usage{
Lorenz(data, target, pred, title = "Lorenz Curve")
}
\arguments{
\item{data}{data.frame consisting of a vector of predicted probabilities and
a vector of actaul (1,0)}

\item{target}{name of response variable}

\item{pred}{name of variable with predicted probabilities}

\item{title}{title of the Lorenz plot. Default is "Lorenz Curve"}
}
\value{
plot of Lorenz curve with C-Stat, KS-Stat, and Gini-Coef
}
\description{
\code{Lorenz} plots the Lorenz curve along with C-Statistic, KS-Statistic,
   and Gini-Coefficient.
}

