% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model and Formula functions.R
\name{Model Functions}
\alias{Model Functions}
\title{Function to help with model building}
\arguments{
\item{variables}{a list or character vector of variables}

\item{target}{the dependent variable to be put before \code{~}}
}
\value{
object of class formula
}
\description{
\code{CreateFormula} creates a formula object for predictive models
\code{Mono} replaces the value in a list used for \code{var.monotone}
\code{Mono_List} creates a list that can be reduced and used for \code{var.monotone}
}

