% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PDP.R
\name{PDP}
\alias{PDP}
\title{Plots Partial Dependency Plots (PDP)}
\usage{
PDP(model, iter, var.name, xlim = NULL, pts = 4000,
  ylab = "P(Conversion)", leg = "right", title = NULL)
}
\arguments{
\item{model}{Model Object (curently only gbm object tested)}

\item{iter}{numberic number of tree to be used from gbm object}

\item{var.name}{independent variable to be plotted}

\item{xlim}{numberic X-axis limit}

\item{pts}{numberic number of points used in graph. Default is \code{4000}}

\item{ylab}{character Y-axis title. Default is \code{"P(Conversion)"}}

\item{leg}{character for legend location. Default \code{"right"}}

\item{title}{chatacer title for plot}
}
\value{
plot of predicted partial dependecy
}
\description{
\code{PDP} plots the Partial Dependency Plots of varaibles in a model object
}

