% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plsmo_plot.R
\name{plsmo_plot}
\alias{plsmo_plot}
\title{Creates a PLSMO Plot}
\usage{
plsmo_plot(mod, data, var.name, target)
}
\arguments{
\item{mod}{model object}

\item{data}{\code{data.frame} model object was build from}

\item{var.name}{name of independent variable}

\item{target}{name of dependent variable}
}
\value{
PLSMO plot
}
\description{
\code{plsmo_plot} creates a PLSMO plot
}

