% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{ROC}
\alias{ROC}
\alias{ROC.default}
\alias{ROC.gam}
\alias{ROC.gbm}
\title{Plots ROC Curve}
\usage{
ROC(model, data, response, ...)

\method{ROC}{default}(model, data, response = NULL, ...)

\method{ROC}{gam}(model, data, response, ...)

\method{ROC}{gbm}(model, data, response = NULL, trees, ...)
}
\arguments{
\item{model}{a model object (e.g. gbm, gam)}

\item{data}{\code{data.frame} of validation data}

\item{response}{depenedent variable name}

\item{...}{all other arugements}
}
\value{
ROC plot
}
\description{
\code{ROC} plots a ROC curve for models with a binary dependent variable
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{gam}: gam objects from mgcv package

\item \code{gbm}: gbm model objects from gbm package
}}

